sorted out de-boost build on linux, fixes #241
